pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-notes-agent"
version = "0.1.0"
description = "AI-powered agent for managing Obsidian notes with LangChain and Claude"
readme = "README.md"
authors = [
    {name = "Duncan", email = "duncan@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.3.0",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.0.20",
    "langgraph>=0.6.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "python-frontmatter>=1.1.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "striprtf>=0.0.26",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
notes = "obsidian_notes_agent.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/notes-agent"
Documentation = "https://github.com/yourusername/notes-agent/blob/main/README.md"
Repository = "https://github.com/yourusername/notes-agent"
"Bug Tracker" = "https://github.com/yourusername/notes-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["obsidian_notes_agent*"]

[tool.setuptools.package-data]
obsidian_notes_agent = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
